refactor(skin): simplify registry generation
