+++ packages/skins/README.md
@@ -9,8 +9,8 @@ The package is private (`"private": true` in `package.json`) and is not publishe
 ## Structure
 
 - `vjsc/` — target-neutral skin components, styles, target transforms, and contract tests.
-- `scripts/` — framework-package materialization and hosted-registry validation.
-- `shadcn/` — Shadcn registry build configuration.
+- `build/` — shared Skin and framework-package output helpers.
+- `registry/` — Shadcn catalog items, build configuration, and focused policy validation.
 - `dev/` — the VJSC React/HTML and CSS/Tailwind development matrix.
 
 ## License
